+++ src/plugins/dexie/rx-storage-instance-dexie.ts
@@ -125,7 +125,7 @@ export class RxStorageInstanceDexie<RxDocType> implements RxStorageInstance<
                             {
                                 _rev: newRevision,
                                 _deleted: insertedIsDeleted,
-                                // TODO attachments are currently not working with lokijs
+                                // TODO attachments are currently not working with dexie.js
                                 _attachments: {} as any
                             }
                         );
@@ -148,6 +148,7 @@ export class RxStorageInstanceDexie<RxDocType> implements RxStorageInstance<
                                 endTime: startTime
                             });
                         }
+
                         ret.success[id] = writeDoc;
                     } else {
                         // update existing document