+++ src/plugins/replication-couchdb.ts
@@ -44,7 +44,7 @@ import type {
     SyncOptions,
     PouchDBInstance
 } from '../types';
-import { transformDocumentDataFromRxStorageToRxDB } from '../rx-storage-helper';
+import { runPluginHooks } from '../hooks';
 
 // add pouchdb-replication-plugin
 addPouchPlugin(PouchReplicationPlugin);
@@ -171,7 +171,17 @@ export function setPouchEventEmitter(
 
                 (ev as any).change.docs
                     .filter((doc: any) => doc.language !== 'query') // remove internal docs
-                    .map((doc: any) => transformDocumentDataFromRxStorageToRxDB(rxRepState.collection, doc)) // do primary-swap and keycompression
+                    .map((doc: any) => {
+                        const hookParams = {
+                            database: rxRepState.collection.database,
+                            primaryPath: rxRepState.collection.schema.primaryPath,
+                            schema: rxRepState.collection.schema.jsonSchema,
+                            doc
+                        };
+
+                        runPluginHooks('postReadFromInstance', hookParams);
+                        return hookParams.doc;
+                    }) // do primary-swap and keycompression
                     .forEach((doc: any) => rxRepState._subjects.docs.next(doc));
             }));
 
@@ -365,27 +375,26 @@ export function syncCouchDB(
     return repState;
 }
 
-export const rxdb = true;
-export const prototypes = {
-    RxCollection: (proto: any) => {
-        proto.syncCouchDB = syncCouchDB;
-    }
-};
 
-export const hooks = {
-    createRxCollection: function (
-        collection: RxCollection
-    ) {
-        const pouch: PouchDBInstance | undefined = collection.storageInstance.internals.pouch;
-        if (pouch) {
-            INTERNAL_POUCHDBS.add(collection.storageInstance.internals.pouch);
-        }
-    }
-};
 
 export const RxDBReplicationCouchDBPlugin: RxPlugin = {
     name: 'replication-couchdb',
-    rxdb,
-    prototypes,
-    hooks
+    rxdb: true,
+    prototypes: {
+        RxCollection: (proto: any) => {
+            proto.syncCouchDB = syncCouchDB;
+        }
+    },
+    hooks: {
+        createRxCollection: {
+            after: function (
+                collection: RxCollection
+            ) {
+                const pouch: PouchDBInstance | undefined = collection.storageInstance.internals.pouch;
+                if (pouch) {
+                    INTERNAL_POUCHDBS.add(collection.storageInstance.internals.pouch);
+                }
+            }
+        }
+    }
 };