+++ test/unit.test.ts
@@ -15,12 +15,13 @@ import './unit/rx-storage-lokijs.test.js';
 import './unit/rx-storage-dexie.test.js';
 import './unit/instance-of-check.test.js';
 import './unit/rx-schema.test.js';
+import './unit/attachments.test.js'; // TODO move down
 import './unit/key-compression.test.js';
 import './unit/bug-report.test.js';
+import './unit/replication-graphql.test.js'; // TODO move down
 import './unit/rx-database.test.js';
 import './unit/rx-collection.test.js';
 import './unit/rx-document.test.js';
-import './unit/attachments.test.js'; // TODO move down
 import './unit/rx-query.test.js';
 import './unit/primary.test.js';
 import './unit/temporary-document.test.js';
@@ -38,14 +39,13 @@ import './unit/hooks.test.js';
 import './unit/orm.test.js';
 import './unit/population.test.js';
 import './unit/backup.test.js';
+import './unit/local-documents.test.js';
 import './unit/leader-election.test.js';
 import './unit/replication.test.js';
-import './unit/replication-graphql.test.js';
 import './unit/replication-couchdb.test.js';
 import './unit/encryption.test.js';
 import './unit/import-export.test.js';
 import './unit/cross-instance.test.js';
 import './unit/server.test.js';
-import './unit/mod-encryption.test.js';
 import './unit/plugin.test.js';
 import './unit/last.test.js';